+++ NEWS.md
@@ -6,6 +6,7 @@ Phan NEWS
 New Features(Analysis):
 - Fix false positive PhanPossiblyUndeclaredVariable warning when a `try` block unconditionally returns/throws/exits (#4419)
 - Fix false positive warnings when analyzing enums, infer that automatically generated methods of enums exist. (#4313)
+- Properly resolve template type when `getIterator` returns an `Iterator` that includes a template. (#4556)
 
 Bug fixes:
 - Fix off-by-one error when inferring from comparison conditions such as `count($arr) > 0` and `count($arr) >= 1` that the array is non-empty. (#4551)